% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutexes.R
\docType{methods}
\name{attach.mutex}
\alias{attach.mutex}
\alias{attach.mutex,character-method}
\alias{attach.mutex,boost.mutex.descriptor-method}
\title{Attach to an existing mutex.}
\usage{
attach.mutex(obj, ...)

\S4method{attach.mutex}{character}(obj, ...)

\S4method{attach.mutex}{boost.mutex.descriptor}(obj, ...)
}
\arguments{
\item{obj}{the descriptor object.}

\item{...}{other arguments needed by attach.}
}
\value{
A mutex.
}
\description{
Attach to an existing mutex using either a file or 
description object
}
