% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutexes.R
\docType{methods}
\name{describe,boost.mutex-method}
\alias{describe,boost.mutex-method}
\title{Describe the boost.mutex object}
\usage{
\S4method{describe}{boost.mutex}(x)
}
\arguments{
\item{x}{the boost mutex object to describe.}
}
\description{
The information required to ``attach'' to an existing
mutex object.
}
