% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutexes.R
\docType{methods}
\name{description}
\alias{description}
\alias{description,descriptor-method}
\title{Accessor for descriptor objects}
\usage{
description(x)

\S4method{description}{descriptor}(x)
}
\arguments{
\item{x}{the descriptor object.}
}
\value{
a list of description information.
}
\description{
Retrieve the list of description information from a 
descriptor object.
}
