% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutexes.R
\docType{class}
\name{descriptor-class}
\alias{descriptor-class}
\title{An S4 class holding mutex description information.}
\description{
Objects of class description allow users to ``attach'' 
to existing mutexes within or across processes.
}
\section{Slots}{

\describe{
\item{\code{description}}{the list of description information.}
}}

