% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutexes.R
\docType{methods}
\name{read}
\alias{read}
\alias{read,boost.mutex-method}
\title{Is it a read (shared) mutex?}
\usage{
read(m)

\S4method{read}{boost.mutex}(m)
}
\arguments{
\item{m}{the mutex}
}
\value{
TRUE if the mutex is read (shared), FALSE otherwise.
}
\description{
Tells the user if a mutex is a read (shared) mutex. If it is 
not then it must be a write (exclusive) mutex.
}

