% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutexes.R
\docType{methods}
\name{shared}
\alias{shared}
\alias{shared,boost.mutex-method}
\title{Is it a shared mutex?}
\usage{
shared(m)

\S4method{shared}{boost.mutex}(m)
}
\arguments{
\item{m}{the mutex}
}
\value{
TRUE if the mutex is shared, FALSE otherwise.
}
\description{
Tells the user if a mutex is a shared mutex. If it is 
not then it must be a write (exclusive) mutex.
}
