\name{synchronicity-package}
\Rdversion{1.1}
\alias{synchronicity-package}
\alias{synchronicity}
\docType{package}
\title{
This package provides support Interprocess Communication (ipc) synchronization
and message passing.
}
\description{
See Description
}
\details{
\tabular{ll}{
Package: \tab synchronicity\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2009-07-12\cr
License: \tab What license is it under?\cr
LazyLoad: \tab yes\cr
}
Use this package for ipc support.
}
\author{
Michael Kane

Maintainer: <kaneplusplus@gmail.com>
}
\references{
http://www.boost.org/doc/libs/1_39_0/doc/html/interprocess.html
}
\examples{
print('synchronicity')
}
