% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutexes.R
\name{shared.name}
\alias{shared.name}
\alias{shared.name,boost.mutex-method}
\title{The name of a mutex's shared resource}
\usage{
shared.name(m)
}
\arguments{
\item{m}{a \code{boost.mutex} object}
}
\value{
A string specifying the shared resource associated with the given 
\code{boost.mutex} object.
}
\description{
This function returns the shared resource associated with a
\code{boost.mutex} object.
}
\examples{
x = boost.mutex()
print(shared.name(x))
}
