\name{synchronicity-package}
\Rdversion{1.0.4}
\alias{synchronicity-package}
\alias{synchronicity}
\docType{package}
\title{
This package provides support for synchronization via mutexes and may eventually
support interprocess communication (ipc) and message passing.
}
\description{
This package provides support for synchronization via mutexes and may eventually
support interprocess communication (ipc) and message passing.
}
\details{
\tabular{ll}{
Package: \tab synchronicity\cr
Type: \tab Package\cr
Version: \tab 1.0.4\cr
Date: \tab 2010-04-10\cr
License: \tab LGPL-3\cr
Copyright: \tab (C) 2010 Michael J. Kane\cr
URL: \tab http://www.bigmemory.org\cr
LazyLoad: \tab yes\cr
}
}
\author{Michael J. Kane <kaneplusplus@gmail.com>}
\references{
The Bigmemory Project: \url{http://www.bigmemory.org/}. //
Boost Interprocess Library: \url{http://www.boost.org/}.
}
\examples{
# No examples are provided here.
}
