\name{synchronicity-package}
\Rdversion{1.0.1}
\alias{synchronicity-package}
\alias{synchronicity}
\docType{package}
\title{
This package provides support for synchronization via mutexes and may eventually
support interprocess communication (ipc) and message passing.
}
\description{
This package provides support for synchronization via mutexes and may eventually
support interprocess communication (ipc) and message passing.
}
\details{
\tabular{ll}{
Package: \tab synchronicity\cr
Type: \tab Package\cr
Version: \tab 1.0.1\cr
Date: \tab 2010-02-22\cr
License: \tab LGPL-3\cr
LazyLoad: \tab yes\cr
}
Use this package for ipc support.
}
\author{
Michael Kane

Maintainer: <kaneplusplus@gmail.com>
}
\references{
Boost Interprocess
}
\examples{
print('synchronicity')
}
