% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uuid.R
\name{uuid}
\alias{uuid}
\title{Create a universal unique identifier.}
\usage{
uuid()
}
\value{
A unique string.
}
\description{
This function creates an identifier that will be (with high 
probability) unique on a single machine or group of machines.
}
\examples{
print(uuid())
}
